mem maze env ok 1
